backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
rembg==2.0.54
basicsr>=1.4.2
realesrgan>=0.3.0
onnxruntime==1.16.3
Pillow==10.2.0
python-dotenv==1.0.0
redis==5.0.1
numpy==1.26.3
opencv-python==4.9.0.80 